+++ lib/ferry/data/providers/schedule/schedule.provider.dart
@@ -15,11 +15,6 @@ abstract class ScheduleProvider with DataLoggy {
     required int endStationId,
     required int dayId,
   });
-
-  Future<Iterable<Map<String, dynamic>>> getEndStations({
-    required int startStationId,
-    required int dayId,
-  });
 }
 
 class IzdenizScheduleProvider extends ScheduleProvider {
@@ -57,30 +52,5 @@ class IzdenizScheduleProvider extends ScheduleProvider {
         .toList();
 
     return hourRows.map((e) => e.text.replaceAll('\\n', '').trim());
-  }
-
-  @override
-  Future<Iterable<Map<String, dynamic>>> getEndStations({
-    required int startStationId,
-    required int dayId,
-  }) async {
-    loggy.debug('Getting end stations...');
-    loggy.trace('start station id: $startStationId');
-    loggy.trace('day id: $dayId');
-
-    final Response<List<dynamic>> response = await _client.post(
-      'https://www.izdeniz.com.tr/tr/IskeleGuncelle',
-      data: {
-        'kalkisIskele': startStationId,
-        'GunTipi': dayId,
-      },
-      options: Options(
-        contentType: Headers.formUrlEncodedContentType,
-        headers: izdenizHeaders,
-      ),
-    );
-
-    final data = List<Map<String, dynamic>>.from(response.data!);
-    return data;
   }
 }